single-sample kernel estimation loop

+++ net.py
@@ -233,7 +233,7 @@ class DecoderAdaConv(nn.Module):
         b, n, h, w = sF['r4_1'].shape
         adaconv_out = {}
         style = self.style_encoding(sF['r4_1'].detach())
-        style, indices, commit_loss = self.vq(style.flatten(2))
+        style, indices, commit_loss = self.vq(style.flatten(2).float())
         style = self.style_projection(style.flatten(1))
         style = style.reshape(b, self.s_d, 4, 4)
         adaconv_out['r4_1'] = self.kernel_1(style, cF['r4_1'])
